reporting_framework/requirements.txt
pandas
openpyxl
python-pptx
rich
questionary
pyyaml
yaspin
